Tighten a bit RoR detection
Phusion passenger is used for Ruby, nodejs, Python
as well as other exotic things, there isn't a 50% chance of Ruby being
used when it's detected.

On the other hand, if `mod_rails` or `mod_rack` is present,
we can be certain that there is indeed Ruby on Rails behind
the website.